pyproject.toml
[project]
name = "wahoo-plan"
version = "0.1.0"
description = "Tools to convert YAML workouts to Wahoo PLAN and plot intervals"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PyYAML>=6.0",
    "matplotlib>=3.5.0",
    "numpy>=1.20.0",
    "pytest>=8.4.1",
    "click>=8.0",
    "rich>=12.0",
]

[project.scripts]
wahoo = "wahoo_plane.cli:cli"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["wahoo_plane"]

[tool.pytest.ini_options]
pythonpath = [
    "."
]
testpaths = [
    "tests"
]
